move Composition into its own module under nucdata it didn't make sense for it to be in rate_collection
