+++ pynucastro/networks/__init__.py
@@ -43,8 +43,7 @@ from .helper import network_helper
 from .network_compare import NetworkCompare
 from .nse_network import NSENetwork
 from .python_network import PythonNetwork
-from .rate_collection import (Composition, Explorer, LoddersComposition,
-                              RateCollection, RateDuplicationError)
+from .rate_collection import Explorer, RateCollection, RateDuplicationError
 from .simple_cxx_network import SimpleCxxNetwork
 from .sympy_network_support import SympyRates
